requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-telegram-bot==20.7
openai==1.10.0
pyyaml==6.0.1
python-dotenv==1.0.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
aiohttp==3.9.1
jinja2==3.1.3
python-multipart==0.0.6
